+++ crates/kernels-metal/kernels/attn/ple.metal
@@ -20,9 +20,9 @@ inline void ple_mask_window(thread int* window, int ngram, int eos) {
 // Engram hashes tokenizer-compressed ids: the window holds raw ids (so the
 // state and the eos barrier stay in the tokenizer's id space) and is mapped
 // after masking, so the pre-start padding maps like every other id.
-inline void ple_map_window(const device int* map, int has_map, thread int* window, int ngram) {
+inline void ple_map_window(const device long* map, int has_map, thread int* window, int ngram) {
   if (!has_map) return;
-  for (int p = 0; p < ngram; ++p) window[p] = map[window[p]];
+  for (int p = 0; p < ngram; ++p) window[p] = int(map[window[p]]);
 }
 
 inline void ple_hash_row(
@@ -58,7 +58,7 @@ inline void ple_hash_row(
     const constant int& heads               [[buffer(6)]],
     const constant int& heads_per_ngram     [[buffer(7)]],
     const constant int& eos                 [[buffer(8)]],
-    const device int* map                   [[buffer(9)]],
+    const device long* map                  [[buffer(9)]],
     const constant int& has_map             [[buffer(10)]],
     uint pos [[thread_position_in_grid]]) {
   const int r = int(pos);
@@ -98,7 +98,7 @@ inline void ple_hash_row(
     const constant int& heads               [[buffer(7)]],
     const constant int& heads_per_ngram     [[buffer(8)]],
     const constant int& eos                 [[buffer(9)]],
-    const device int* map                   [[buffer(10)]],
+    const device long* map                  [[buffer(10)]],
     const constant int& has_map             [[buffer(11)]],
     uint pos [[thread_position_in_grid]]) {
   const int r = int(pos);
@@ -156,7 +156,7 @@ inline void ple_hash_row(
     const constant int& heads               [[buffer(10)]],
     const constant int& heads_per_ngram     [[buffer(11)]],
     const constant int& eos                 [[buffer(12)]],
-    const device int* map                   [[buffer(13)]],
+    const device long* map                  [[buffer(13)]],
     const constant int& has_map             [[buffer(14)]],
     uint pos [[thread_position_in_grid]]) {
   const int r = int(pos);
